Parsing HTTP using picohttpparser C library
Working example of calling a statically linked C library for parsing
HTTP requests. Note all of the @ptrCast's in order to cast to. Zig seems
to think all pointers to/from C are optional. This is likely because Zig
can't verify what sort of mean things the C code might be doing to it.